+++ products/jbrowse-desktop/src/StartScreen/LauncherPanel.tsx
@@ -1,16 +1,7 @@
 import React, { useState } from 'react'
-import {
-  Button,
-  FormControl,
-  Grid,
-  InputLabel,
-  MenuItem,
-  Select,
-  makeStyles,
-} from '@material-ui/core'
+import { Button, Grid, makeStyles } from '@material-ui/core'
 import PluginManager from '@jbrowse/core/PluginManager'
-import { createPluginManager } from './util'
-import preloadedConfigs from './preloadedConfigs'
+import PreloadedDatasetSelector from './PreloadedDatasetSelector'
 import OpenSequenceDialog from './OpenSequenceDialog'
 import OpenJBrowseWebConfigDialog from './OpenJBrowseWebConfigDialog'
 
@@ -57,7 +48,7 @@ export default function StartScreenOptionsPanel({
           </Button>
         </Grid>
         <Grid item>
-          <PreloadedSelector setPluginManager={setPluginManager} />
+          <PreloadedDatasetSelector setPluginManager={setPluginManager} />
         </Grid>
       </Grid>
       {sequenceDialogOpen ? (
@@ -73,47 +64,5 @@ export default function StartScreenOptionsPanel({
         />
       ) : null}
     </Grid>
-  )
-}
-
-function PreloadedSelector({
-  setPluginManager,
-}: {
-  setPluginManager: (arg0: PluginManager) => void
-}) {
-  const classes = useStyles()
-  const [assemblyChoice, setAssemblyChoice] = useState('hg38')
-  return (
-    <div style={{ display: 'flex' }}>
-      <FormControl className={classes.formControl}>
-        <InputLabel shrink>Pre-loaded datasets</InputLabel>
-        <Select
-          value={assemblyChoice}
-          label="Pre-loaded datasets"
-          onChange={event => setAssemblyChoice(event.target.value as string)}
-        >
-          {Object.keys(preloadedConfigs).map(name => (
-            <MenuItem key={name} value={name}>
-              {name}
-            </MenuItem>
-          ))}
-        </Select>
-      </FormControl>
-      <div style={{ margin: 'auto 0' }}>
-        <Button
-          onClick={async () => {
-            const pm = await createPluginManager(
-              // @ts-ignore
-              preloadedConfigs[assemblyChoice],
-            )
-            setPluginManager(pm)
-          }}
-          variant="contained"
-          color="primary"
-        >
-          Go
-        </Button>
-      </div>
-    </div>
   )
 }
